move the focus trapping to its own module
